Azure/Functions/KeywordCollector/requirements.txt
azure-functions
pyodbc
requests
python-dotenv
google-ads
PyYAML
